+++ espnet/asr/asr_pytorch.py
@@ -22,28 +22,28 @@ from chainer.training import extensions
 import torch
 
 # espnet related
-from asr_utils import adadelta_eps_decay
-from asr_utils import add_results_to_json
-from asr_utils import CompareValueTrigger
-from asr_utils import get_model_conf
-from asr_utils import load_inputs_and_targets
-from asr_utils import make_batchset
-from asr_utils import PlotAttentionReport
-from asr_utils import restore_snapshot
-from asr_utils import torch_load
-from asr_utils import torch_resume
-from asr_utils import torch_save
-from asr_utils import torch_snapshot
-from e2e_asr_th import E2E
-from e2e_asr_th import Loss
-from e2e_asr_th import pad_list
+from espnet.asr.asr_utils import adadelta_eps_decay
+from espnet.asr.asr_utils import add_results_to_json
+from espnet.asr.asr_utils import CompareValueTrigger
+from espnet.asr.asr_utils import get_model_conf
+from espnet.asr.asr_utils import load_inputs_and_targets
+from espnet.asr.asr_utils import make_batchset
+from espnet.asr.asr_utils import PlotAttentionReport
+from espnet.asr.asr_utils import restore_snapshot
+from espnet.asr.asr_utils import torch_load
+from espnet.asr.asr_utils import torch_resume
+from espnet.asr.asr_utils import torch_save
+from espnet.asr.asr_utils import torch_snapshot
+from espnet.nets.e2e_asr_th import E2E
+from espnet.nets.e2e_asr_th import Loss
+from espnet.nets.e2e_asr_th import pad_list
 
 # for kaldi io
 import kaldi_io_py
 
 # rnnlm
-import extlm_pytorch
-import lm_pytorch
+import espnet.lm.extlm_pytorch as extlm_pytorch
+import espnet.lm.lm_pytorch as lm_pytorch
 
 # matplotlib related
 import matplotlib
