+++ espnet/lm/lm_chainer.py
@@ -28,13 +28,13 @@ from chainer import reporter
 from chainer import training
 from chainer.training import extensions
 
-from lm_utils import compute_perplexity
-from lm_utils import count_tokens
-from lm_utils import MakeSymlinkToBestModel
-from lm_utils import ParallelSentenceIterator
-from lm_utils import read_tokens
+from espnet.lm.lm_utils import compute_perplexity
+from espnet.lm.lm_utils import count_tokens
+from espnet.lm.lm_utils import MakeSymlinkToBestModel
+from espnet.lm.lm_utils import ParallelSentenceIterator
+from espnet.lm.lm_utils import read_tokens
 
-import deterministic_embed_id as DL
+import espnet.nets.deterministic_embed_id as DL
 
 REPORT_INTERVAL = 100
