+++ espnet/nets/deterministic_embed_id.py
